Story 1.5, review: an executed Action is a wait, a window is a wait from its second frame, the resurrection warning counts
The fairness review passed on parity and found the confirmation rule short in three ways. An act
can begin ready and end ready in one go and announce nothing: a move the hero cannot take, or the
first floor's entrance refused without the amulet, which posts a message; the notification and
the window's change both miss the wait that follows, and AD-5 gives every executed Action its own
wait, so an Action handed to the game, seen as the hero holding it when stepping resumes, is now
the third source of a new wait. The chasm prompt refuses an answer until it has been updated for
more than 0.2 s of frame time, so a window confirmed on the frame it appeared could not be
answered; a window is a wait from its second frame in front. And the resurrection window stacks a
warning over itself when a kept-item slot is empty, which the confirmation now recognises like any
Prompt; the inventory pane selecting an item, with no window and the map refusing clicks, is not a
wait and is named in the diagnostic.

Also from the review: k is a long counting confirmed waits from 1, as ADR-0013 has it, with the
spine's "0-based" read as the value before the first wait; the boot sets the support-prompt flag
so the first boss's key opens no window; the Prompts comment says what the list is and is not;
the click-path citation is the default cell listener, not the selector; the ledger and the
amendment count the hook's nine added lines; the wake-up test asserts exactly sixty parks; four
tests added (the refused move, the refused ascent held until its message is dismissed, a full
rest, the resurrection warning).

+++ shatterfish/harness/src/main/java/org/shatterfish/harness/driver/WaitSequence.java
@@ -18,24 +18,24 @@ package org.shatterfish.harness.driver;
 public interface WaitSequence<O, D> {
 
     /** Reseeds the game's generator for wait {@code k}; story 1.15. */
-    default void reseed(int k) {
+    default void reseed(long k) {
     }
 
     /** Reads what the player could see at wait {@code k}; the Observer stories. */
-    default O observe(int k) {
+    default O observe(long k) {
         return null;
     }
 
     /** Decides what to do at wait {@code k} from what was observed; the brain, or an agent. */
-    default D decide(int k, O observation) {
+    default D decide(long k, O observation) {
         return null;
     }
 
     /** Does what was decided at wait {@code k}, the way the player would; story 1.13. */
-    default void execute(int k, D decision) {
+    default void execute(long k, D decision) {
     }
 
     /** Writes the record of wait {@code k}; E3. */
-    default void record(int k, O observation, D decision) {
+    default void record(long k, O observation, D decision) {
     }
 }
